# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bilingual personal website + CMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  id: number;
  username: string;
  isDefaultPassword: boolean;
}

export interface DashboardStats {
  totalPages: number;
  publishedPages: number;
  draftPages: number;
  totalSections: number;
}

export interface Page {
  id: number;
  titleAr: string;
  titleEn: string;
  slug: string;
  isPublished: boolean;
  showInNav: boolean;
  isHomepage: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PageInput {
  /** @minLength 1 */
  titleAr: string;
  /** @minLength 1 */
  titleEn: string;
  /** @minLength 1 */
  slug: string;
  isPublished?: boolean;
  showInNav?: boolean;
}

export interface PageUpdate {
  /** @minLength 1 */
  titleAr?: string;
  /** @minLength 1 */
  titleEn?: string;
  /** @minLength 1 */
  slug?: string;
  isPublished?: boolean;
  showInNav?: boolean;
}

export type SectionWithImagesType = typeof SectionWithImagesType[keyof typeof SectionWithImagesType];


export const SectionWithImagesType = {
  hero: 'hero',
  text: 'text',
  text_with_image: 'text_with_image',
  image_gallery: 'image_gallery',
  cards_grid: 'cards_grid',
  timeline: 'timeline',
  contact_strip: 'contact_strip',
} as const;

export interface BulletItem {
  id: string;
  textAr?: string;
  textEn?: string;
}

export interface SectionDataItem {
  id: string;
  titleAr?: string;
  titleEn?: string;
  descriptionAr?: string;
  descriptionEn?: string;
  subheadingAr?: string;
  subheadingEn?: string;
  date?: string;
  icon?: string;
  bullets?: BulletItem[];
}

export interface SectionData {
  titleAr?: string;
  titleEn?: string;
  contentAr?: string;
  contentEn?: string;
  imagePosition?: string;
  email?: string;
  linkedin?: string;
  locationAr?: string;
  locationEn?: string;
  items?: SectionDataItem[];
}

export interface SectionImage {
  id: number;
  sectionId: number;
  filename: string;
  captionAr: string;
  captionEn: string;
  sortOrder: number;
  url: string;
}

export interface SectionWithImages {
  id: number;
  pageId: number;
  type: SectionWithImagesType;
  sortOrder: number;
  isVisible: boolean;
  data: SectionData;
  images: SectionImage[];
}

export interface PageWithSections {
  id: number;
  titleAr: string;
  titleEn: string;
  slug: string;
  isPublished: boolean;
  showInNav: boolean;
  isHomepage: boolean;
  createdAt: string;
  updatedAt: string;
  sections: SectionWithImages[];
}

export type SectionType = typeof SectionType[keyof typeof SectionType];


export const SectionType = {
  hero: 'hero',
  text: 'text',
  text_with_image: 'text_with_image',
  image_gallery: 'image_gallery',
  cards_grid: 'cards_grid',
  timeline: 'timeline',
  contact_strip: 'contact_strip',
} as const;

export interface Section {
  id: number;
  pageId: number;
  type: SectionType;
  sortOrder: number;
  isVisible: boolean;
  data: SectionData;
}

export type SectionInputType = typeof SectionInputType[keyof typeof SectionInputType];


export const SectionInputType = {
  hero: 'hero',
  text: 'text',
  text_with_image: 'text_with_image',
  image_gallery: 'image_gallery',
  cards_grid: 'cards_grid',
  timeline: 'timeline',
  contact_strip: 'contact_strip',
} as const;

export interface SectionInput {
  type: SectionInputType;
  data: SectionData;
}

export interface SectionUpdate {
  data: SectionData;
}

export interface VisibilityUpdate {
  isVisible: boolean;
}

export interface ReorderRequest {
  ids: number[];
}

export interface ImageUpdate {
  captionAr?: string;
  captionEn?: string;
}

export interface PasswordChange {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface UsernameChange {
  /** @minLength 3 */
  newUsername: string;
  currentPassword: string;
}

export interface NavItem {
  id: number;
  titleAr: string;
  titleEn: string;
  slug: string;
  isHomepage: boolean;
}

export interface PublicPage {
  id: number;
  titleAr: string;
  titleEn: string;
  slug: string;
  isHomepage: boolean;
  sections: SectionWithImages[];
}