# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bilingual personal website + CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/session:
    get:
      operationId: getAdminSession
      tags: [auth]
      summary: Check current admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all pages
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "201":
          description: Page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /pages/{id}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get page with its sections
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page with sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWithSections"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update page metadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageUpdate"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page (non-homepage only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Cannot delete homepage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Sections ──────────────────────────────────────────────────────────────
...
  /images/{id}:
    patch:
      operationId: updateImage
      tags: [images]
      summary: Update image captions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUpdate"
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete an image and its file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /sections/{sectionId}/images/reorder:
    post:
      operationId: reorderImages
      tags: [images]
      summary: Reorder images in a section
      parameters:
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Account ───────────────────────────────────────────────────────────────
  /account/password:
    patch:
      operationId: changePassword
      tags: [account]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /account/username:
    patch:
      operationId: changeUsername
      tags: [account]
      summary: Change admin username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsernameChange"
      responses:
        "200":
          description: Username changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Public endpoints ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────
    LoginCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [id, username, isDefaultPassword]
      properties:
        id:
          type: integer
        username:
          type: string
        isDefaultPassword:
          type: boolean

    # ── Dashboard ─────────────────────────────────────────────────────────
...
    Page:
      type: object
      required: [id, titleAr, titleEn, slug, isPublished, showInNav, isHomepage, createdAt, updatedAt]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        titleEn:
          type: string
        slug:
          type: string
        isPublished:
          type: boolean
        showInNav:
          type: boolean
        isHomepage:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PageInput:
      type: object
      required: [titleAr, titleEn, slug]
      properties:
        titleAr:
          type: string
          minLength: 1
        titleEn:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        isPublished:
          type: boolean
        showInNav:
          type: boolean
    PageUpdate:
      type: object
      properties:
        titleAr:
          type: string
          minLength: 1
        titleEn:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        isPublished:
          type: boolean
        showInNav:
          type: boolean
    PageWithSections:
      type: object
      required: [id, titleAr, titleEn, slug, isPublished, showInNav, isHomepage, createdAt, updatedAt, sections]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        titleEn:
          type: string
        slug:
          type: string
        isPublished:
          type: boolean
        showInNav:
          type: boolean
        isHomepage:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionWithImages"

    # ── Section Data (structured to avoid looseObject generation) ─────────
    # This covers the union of all section type fields.
    # The server stores the full JSON; this schema validates known fields.
    SectionDataItem:
      type: object
      required: [id]
      properties:
        id:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        subheadingAr:
          type: string
        subheadingEn:
          type: string
        date:
          type: string
        icon:
          type: string
        bullets:
          type: array
          items:
            $ref: "#/components/schemas/BulletItem"
    BulletItem:
      type: object
      required: [id]
      properties:
        id:
          type: string
        textAr:
          type: string
        textEn:
          type: string
    SectionData:
      type: object
      properties:
        titleAr:
          type: string
        titleEn:
          type: string
        contentAr:
          type: string
        contentEn:
          type: string
        imagePosition:
          type: string
        email:
          type: string
        linkedin:
          type: string
        locationAr:
          type: string
        locationEn:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SectionDataItem"

    # ── Sections ──────────────────────────────────────────────────────────
...
    SectionWithImages:
      type: object
      required: [id, pageId, type, sortOrder, isVisible, data, images]
      properties:
        id:
          type: integer
        pageId:
          type: integer
        type:
          type: string
          enum: [hero, text, text_with_image, image_gallery, cards_grid, timeline, contact_strip]
        sortOrder:
          type: integer
        isVisible:
          type: boolean
        data:
          $ref: "#/components/schemas/SectionData"
        images:
          type: array
          items:
            $ref: "#/components/schemas/SectionImage"
...
    ReorderRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer

    # ── Images ────────────────────────────────────────────────────────────
    SectionImage:
      type: object
      required: [id, sectionId, filename, captionAr, captionEn, sortOrder, url]
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        filename:
          type: string
        captionAr:
          type: string
        captionEn:
          type: string
        sortOrder:
          type: integer
        url:
          type: string
    ImageUpdate:
      type: object
      properties:
        captionAr:
          type: string
        captionEn:
          type: string

    # ── Account ───────────────────────────────────────────────────────────
    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    UsernameChange:
      type: object
      required: [newUsername, currentPassword]
      properties:
        newUsername:
          type: string
          minLength: 3
        currentPassword:
          type: string

    # ── Public ────────────────────────────────────────────────────────────
...