# lib/api-zod/src/generated/types/pageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bilingual personal website + CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PageInput {
  /** @minLength 1 */
  titleAr: string;
  /** @minLength 1 */
  titleEn: string;
  /** @minLength 1 */
  slug: string;
  isPublished?: boolean;
  showInNav?: boolean;
}