# lib/api-zod/src/generated/types/sectionImage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bilingual personal website + CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SectionImage {
  id: number;
  sectionId: number;
  filename: string;
  captionAr: string;
  captionEn: string;
  sortOrder: number;
  url: string;
}