# lib/api-zod/src/generated/types/sectionWithImages.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Bilingual personal website + CMS API
 * OpenAPI spec version: 0.1.0
 */
import type { SectionData } from './sectionData';
import type { SectionImage } from './sectionImage';
import type { SectionWithImagesType } from './sectionWithImagesType';

export interface SectionWithImages {
  id: number;
  pageId: number;
  type: SectionWithImagesType;
  sortOrder: number;
  isVisible: boolean;
  data: SectionData;
  images: SectionImage[];
}